==> code/02_Data_Cleaning.ipynb <==
# Subreddit Classification: Data Cleaning

# Imports
import pandas as pd

# Read in data
df = pd.read_csv('../data/conversations.csv')

df.head(3)

df.shape

# df.info()
df.info()

type(df['comts_first'][0])

april = []
for i, title in enumerate(df['title']):
    if 'april fool' in title.lower():
        april.append(i)
print('Number of April Fools Titles: ', len(april))

for i in april:
    print(df.iloc[i]['title'])

### Investigating Nulls

# How to deal with nulls
df.isnull().sum()[df.isnull().sum() != 0]

# Look at rows with two nulls for 'selftext'
df[df['selftext'].isnull()]

# Two have values of reddit.come for domain
df['domain'].value_counts()

# The two posts from reddit.com are promoting the sister sub...remove theses rows
print(df[(df['selftext'].isnull()) & (df['domain'] == 'reddit.com')]['title'].values[0])
print(df[(df['selftext'].isnull()) & (df['domain'] == 'reddit.com')]['title'].values[1])

# Remove the two rows where 'reddit.com' is the domain name
df.drop(labels=df[df['domain'] == 'reddit.com'].index, inplace=True)

# Reset index.
df.reset_index(drop=True, inplace=True)

# Verify changes
df['domain'].value_counts()

# Fill remaining empty selftext row with an empty string
df[df['selftext'].isnull()]

df['selftext'].fillna('', inplace=True)

# Verify 'selftext' is no longer null (the only column that had nulls)
df['selftext'].isnull().sum()

### Look for duplicates

# No duplicates
df[df.duplicated()==True]

### Check values in the columns

df.columns

df['subreddit'].value_counts()

df['domain'].value_counts()
# Essentially the same as subreddit so can remove

df['is_original_content'].value_counts()
# Drop...all are false

df['is_video'].value_counts()
# Drop...all are false

df['media_only'].value_counts()
# Drop...all are false

# Drop above columns
df.drop(columns=['domain', 'is_original_content', 'is_video', 'media_only'], inplace=True)

# Verify 4 columns were dropped (originally 18)
df.shape

### Remove rows containing the subreddit name

# Look at rows containing 'CasualConversation'
for i, title in enumerate(df['title']):
    if 'casualconversation' in title.lower():
        print(i, title)

# Look at rows containing 'SeriousConversation'
for i, title in enumerate(df['title']):
    if 'seriousconversation' in title.lower():
        print(i, title)

# Get index for rows to remove
remove_casual = [i for i, title in enumerate(df['title']) if 'casualconversation' in title.lower()]
remove_serious = [i for i, title in enumerate(df['title']) if 'seriousconversation' in title.lower()]
remove = remove_casual + remove_serious
remove

# Drop rows from list (and reset index)
# Why? These are not a large percent of the data and I do not beleive they will add any benefit by keeping them
# Rows = 1727, Removing = 6 (6/1727 = 0.3%)
df.drop(labels=remove, inplace=True)
df.reset_index(drop=True, inplace=True)

# Verify dropped rows (1727 - 6 = 1721)
df.shape

### Remove rows where the author is the moderator

# Check for rows where the author is 'AutoModerator'
df[df['author'] == 'AutoModerator']['subreddit'].value_counts()

# What do these posts look like
df[df['author'] == 'AutoModerator'][['subreddit', 'title', 'selftext', 'comts_first']]

# Get index for these rows and drop rows (and reset index)
# Why? Comments may be helpful, but the title and selftext do not offer anything that could help
# There are more from SeriousConversation anyways and we have more observations from SeriousConversations 
# so okay to drop?
moderator_rows = df[df['author'] == 'AutoModerator'].index
df.drop(labels=moderator_rows, inplace=True)
df.reset_index(drop=True, inplace=True)

# Get new value counts for each subreddit
df['subreddit'].value_counts()

### Clean up text columns

df[['title', 'selftext', 'comts_first', 'comts_second']].head()

# Import RegEx
import re

# Reference: Code adapted from NLP_EDA-InClass in DEN Flex by Sam Stack

# Create function that takes in a string and returns a string with all lowercase and
# Removes punctionation (including underscrose)
def clean_text(raw_text):
    words = re.sub(r'[^a-z0-9]', r' ', raw_text.lower()).split()
    return ' '.join(words)

# Apply function to text columns
for column in ['title', 'selftext', 'comts_first', 'comts_second']:
    df[column] = df[column].apply(clean_text)

df[['title', 'selftext', 'comts_first', 'comts_second']].head()

### Make One Big Text Column

# Combine all text columns
df['text'] = df['title'] + df['selftext'] + df['comts_first'] + df['comts_second']

df['text'][0]

### Create Target Column

# Create target column
df['is_serious'] = df['subreddit'].map(lambda x: 1 if x == 'SeriousConversation' else 0)

df['is_serious'].value_counts()

### Save

df.shape

# Save to new csv data folder
df.to_csv('../data/conversations_clean.csv', index = False)

### Clean Comments Datframe

# Read in data
coms_df = pd.read_csv('../data/comments.csv')

coms_df.head(3)

# The subreddit column contains the permalink which contains the subreddit name
coms_df['subreddit'].value_counts().head()

# Map permalink to target column (is_serious)
coms_df['is_serious'] = coms_df['subreddit'].map(lambda x: 1 if 'serious' in x.lower() else 0)
coms_df['is_serious'].value_counts()

# drop subreddit colimn with permalink...no longer needed
coms_df.drop(columns=['subreddit'], inplace=True)

# Clean comments column
print(coms_df['comment'][0])
coms_df['comment'] = coms_df['comment'].apply(clean_text)
coms_df['comment'] = coms_df['comment'].map(lambda x: x.strip())
print(coms_df['comment'][0])

# Check for any nulls
coms_df.isnull().sum()

# Remove comments that are just blank strings
coms_df[coms_df['comment'] == ' ']

coms_df[coms_df['comment'] == '   ']

# Drop these rows
coms_df.drop(labels=[1200, 3121, 16489, 2416, 11253], inplace=True)

coms_df.shape

# Save to new csv data folder
coms_df.to_csv('../data/comments_clean.csv', index = False)
