==> code/06_Model_Selection.ipynb <==
import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier
from sklearn.model_selection import GridSearchCV, train_test_split, cross_val_score
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression 
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.metrics import confusion_matrix

### Read in the Data

# Import data
cv_df = pd.read_csv('../data/count_vec.csv')
tfidf_df = pd.read_csv('../data/tfidf.csv')

cv_df.head()

cv_df.shape

tfidf_df.head()

tfidf_df.shape

### Evaluate Various Models

#### Function: Confusion Matrix Results

def confusion_matrix_results(y_test, predictions):
    # Create confustion matrix and conver to dataframe
    cm = confusion_matrix(y_test, predictions)
    print('-------------CONFUSION MATRIX---------------')
    print(pd.DataFrame(cm, 
                       columns=['pred neg', 'pred pos'], 
                       index = ['actual neg', 'actual pos']))
    
    # Results
    tn, fp, fn, tp = confusion_matrix(y_test, predictions).ravel()
    print("\nTrue Negatives: %s" % tn)
    print("False Positives: %s" % fp)
    print("False Negatives: %s" % fn)
    print("True Positives: %s" % tp)
    print('------------------METRICS-------------------')
    print(f'Sensativity: {round(tp / (tp + fn),4)}')
    print(f'Specificity: {round(tn / (tn + fp),4)}\n\n')
    # Other metrics?? missclassification? 

#### LOGISTIC REGRESSION

def logisitic_reg(df):
    print('*********** LOGISTIC REGRESSION MODEL ************\n')
    features = [column for column in df.columns if column != 'is_serious']
    features.remove('sent_compound')
    features.remove('sent_neu')
    features.remove('char_count') # giving better scores
    X = df[features]
    y = df['is_serious']
    
    # Train Test Split
    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42, stratify = y)
    
    # Instantiate Logistic Regression model
    lr = LogisticRegression(solver='lbfgs')
    
    # Fit the model
    lr.fit(X_train, y_train)
    
    # Score the model
    print('--------------MODEL EVALUTATION---------------')
    print('Cross Val Score: ', cross_val_score(lr, X_train, y_train, cv = 5).mean())
    print('Train Score:     ', lr.score(X_train, y_train))
    print('Test Score:      ', lr.score(X_test, y_test))
    
    # Get predictions
    predictions = lr.predict(X_test)
    
    # Confusion matrix and measures
    confusion_matrix_results(y_test, predictions)

# Tf-Idf
logisitic_reg(tfidf_df)

# Count Vec
logisitic_reg(cv_df)

Observations:
- Best Cross Val Score: 0.85 -- Logistic Regression with Tf-Idf
- Using Tf-Idf leads to a better score (higher cross val, higher sensativity and higher sepcificity) than Count Vectorize
- The Tf-Idf Logistic Regression model is still overfit (train = 0.94, Test = 0.813)

#### KNN

def knn(df):
    print('*********** KNN MODEL ************\n')
    features = [column for column in df.columns if column != 'is_serious']
    features.remove('sent_compound')
    features.remove('sent_neu')
    features.remove('char_count')
    
    X = df[features]
    y = df['is_serious']
    
    # Train Test Split
    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42, stratify = y)

    # Instantiate KNN Model
    knn = KNeighborsClassifier()
    
    # Fit the model
    knn.fit(X_train, y_train)
    
    # Score the model
    print('--------------MODEL EVALUTATION---------------')
    print('Cross Val Score: ', cross_val_score(knn, X_train, y_train, cv = 5).mean())
    print('Train Score:     ', knn.score(X_train, y_train))
    print('Test Score:      ', knn.score(X_test, y_test))
    
    # Get predictions
    predictions = knn.predict(X_test)
    
    # Confusion matrix and measures
    confusion_matrix_results(y_test, predictions)

# Tf-Idf
knn(tfidf_df)

# Count Vec
knn(cv_df)

Observations:
- Best Cross Val Score: 0.73 -- KNN with Tf-Idf
- Tf-Idf performs better thant Count Vec (based on cross val score)
- There is a large difference in the Sensativity metric...tf-idf: 0.889, cross vec: 0.404
    - Tf-Idf really helps reduce false negatives (134 to 25)
- There is a large difference in the Specificity metric...tf-idf: 0.532, cross vec: 0.949
    - Cross Vec really helps reduce false positive (93 to 10)
- The Tf-Idk model us less overfit (still overfit though) than the Cross Vec model

#### DECISION TREE

def decision_tree(df):
    print('*********** DECISION TREE MODEL ************\n')
    features = [column for column in df.columns if column != 'is_serious']
    features.remove('sent_compound')
    features.remove('sent_neu')
    features.remove('char_count')
    X = df[features]
    y = df['is_serious']
    
    # Train Test Split
    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42, stratify = y)
    
    # Instantiate DecisionTree Model
    dt = DecisionTreeClassifier(random_state=45,
                                max_depth = 7, 
                                min_samples_split = 15, 
                                min_samples_leaf = 6)
    # Fit the model
    dt.fit(X_train, y_train)
    
    # Score the model/Evaluate Model
    print('--------------MODEL EVALUTATION---------------')
    print('Cross Val Score: ', cross_val_score(dt, X_train, y_train, cv = 5).mean())
    print('Train Score:     ', dt.score(X_train, y_train))
    print('Test Score:      ', dt.score(X_test, y_test))
 
    # Get predictions
    predictions = dt.predict(X_test)
    
    # Confusion matrix and measures
    confusion_matrix_results(y_test, predictions)

# Tf-Idf
decision_tree(tfidf_df)

# Count Vec
decision_tree(cv_df)

Observations:
- Best Cross Val Score: 0.71 -- Decision Tree with Cross Vec
- Cross performs better thant Tf-Idf (based on cross val score)...unlike the previous two models
- Tf-Idf improves sensativity while Count Vec improves sepcificity (like above)
- Tf-Idf has a much more evenly distributed number of false positive and false negatives than Cross Vec

#### RANDOM FOREST

def random_forest(df):
    print('*********** RANDOM FOREST MODEL ************\n')
    features = [column for column in df.columns if column != 'is_serious']
    features.remove('sent_compound')
    features.remove('sent_neu')
    features.remove('char_count')
    X = df[features]
    y = df['is_serious']
    
    # Train Test Split
    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42, stratify = y)
    
    # Instantiate KNN Model
    rf = RandomForestClassifier(n_estimators=10, 
                                max_depth = None, 
                                max_features = 'auto')
    
    # Fit the model
    rf.fit(X_train, y_train)
    
    # Score the model
    print('--------------MODEL EVALUTATION---------------')
    print('Cross Val Score: ', cross_val_score(rf, X_train, y_train, cv = 5).mean())
    print('Train Score:     ', rf.score(X_train, y_train))
    print('Test Score:      ', rf.score(X_test, y_test))
    
    # Get predictions
    predictions = rf.predict(X_test)
    
    # Confusion matrix and measures
    confusion_matrix_results(y_test, predictions)

# Tf-Idf
random_forest(tfidf_df)

# Count Vec
random_forest(cv_df)

Observations:
- Best Cross Val Score: 0.76 -- Random Forest with Cross Vec
- Cross performs better thant Tf-Idf (based on cross val score)
- Both models have a specificity metric around 82-83%
- The Corss Vec model has a slighly better senstitiviy metric (less false negatives)

#### EXTRA TREE

def extra_trees(df):
    print('*********** EXTRA TREES MODEL ************\n')
    features = [column for column in df.columns if column != 'is_serious']
    features.remove('sent_compound')
    features.remove('sent_neu')
    features.remove('char_count')
    X = df[features]
    y = df['is_serious']
    
    # Train Test Split
    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42, stratify = y)
    
    # Instantiate KNN Model
    rf = RandomForestClassifier(n_estimators=10, 
                                max_depth = None, 
                                max_features = 'auto')
    
    # Train Test Split
    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42, stratify = y)
    
    # Instantiate KNN Model
    et = ExtraTreesClassifier(n_estimators=10)
    
    # Fit the model
    et.fit(X_train, y_train)
    
    # Score the model
    print('--------------MODEL EVALUTATION---------------')
    print('Cross Val Score: ', cross_val_score(et, X_train, y_train, cv = 5).mean())
    print('Train Score:     ', et.score(X_train, y_train))
    print('Test Score:      ', et.score(X_test, y_test))
    
    # Get predictions
    predictions = et.predict(X_test)
    
    # Confusion matrix and measures
    confusion_matrix_results(y_test, predictions)

# Tf-idf
extra_trees(tfidf_df)

# Count Vec
extra_trees(cv_df)

Observations:
- Best Cross Val Score: 0.759 -- Extra Tress with Tf-Idf
- Both models have very similar cross val scores (0.75)
- The Tf-Idf does a better job predicting less false positives

### Top 3 Models
1. Logistic Regression uing Tf-Idf (cv: 0.8504352952160914)
2. Logistic Regression uing Cross Vectorization (cv: 0.8291595984819498)
3. Random Forest using Cross Vectorization (cv: 0.7622948456716102)

### Grid Search

#### Logistic Regression using Tf-Idf

# Set X and y
features = [column for column in tfidf_df.columns if column != 'is_serious']
features.remove('sent_compound')
features.remove('sent_neu')
features.remove('char_count')
X = tfidf_df[features]
y = tfidf_df['is_serious']

# Train Test Split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42, stratify = y)

# Gridsearch for Logistic Regression
lr = LogisticRegression()
lr_params = {
    'solver': ['lbfgs', 'liblinear'],
    'C' : [0.1,1,10,100],
    'max_iter': [150]
}
gs = GridSearchCV(lr, param_grid=lr_params, cv=3)

gs.fit(X_train, y_train)

# Scores
print('Best Cross Val: ', gs.best_score_)
print('Best Params: ', gs.best_params_)
print('Train w/ Params: ', gs.score(X_train, y_train))
print('Test w/ Params: ', gs.score(X_test, y_test))

Obsevation: 0.1 improvement from CV before gridsearch, still overfit

#### Random Forest using Cross Vectorization

# Set X and y
features = [column for column in cv_df.columns if column != 'is_serious']
features.remove('sent_compound')
features.remove('sent_neu')
features.remove('char_count')
X = cv_df[features]
y = cv_df['is_serious']

# Train Test Split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42, stratify = y)

# Gridsearch for Random Forest
rf = RandomForestClassifier()
rf_params = {
    'n_estimators': [50, 100, 150],
    'max_depth' : [10, 20, 30],
    'max_features': ['auto', 1000, 2000]
}
gs = GridSearchCV(rf, param_grid=rf_params, cv=3) 

gs.fit(X_train, y_train)

# Scores
print('Best Cross Val: ', gs.best_score_)
print('Best Params: ', gs.best_params_)
print('Train w/ Params: ', gs.score(X_train, y_train))
print('Test w/ Params: ', gs.score(X_test, y_test))

Observations: Improved CV from 0.76 to 0.81 (but still equally as overfit)

#### Best Model: Logistic Regression with Tf-Idf and Grid Search parameters

### AdaBoost

# Import AdaBoost
from sklearn.ensemble import AdaBoostClassifier

# Set X and y
features = [column for column in tfidf_df.columns if column != 'is_serious']
features.remove('sent_compound')
features.remove('sent_neu')
features.remove('char_count')
X = tfidf_df[features]
y = tfidf_df['is_serious']

# Train Test Split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42, stratify = y)

# Logistic Regression with Tf-Idf and Grid Search Paramters
ada = AdaBoostClassifier(base_estimator=LogisticRegression(C = 10,
                                                           max_iter = 150,
                                                           solver = 'liblinear'))
ada_params = {'n_estimators': [60, 70, 80]}
                         
gs = GridSearchCV(ada, param_grid=ada_params, cv=3)
                         
gs.fit(X_train, y_train);

                         # Scores
print('Best Cross Val: ', gs.best_score_)
print('Best Params: ', gs.best_params_)
print('Train w/ Params: ', gs.score(X_train, y_train))
print('Test w/ Params: ', gs.score(X_test, y_test))

Observation: AdaBoost did NOT improve my cross val score

### SVD

# Import SVD
from sklearn.decomposition import TruncatedSVD

# Get feature matrix
features = [column for column in tfidf_df.columns if column != 'is_serious']
features.remove('sent_compound')
features.remove('sent_neu')
features.remove('char_count')
X = tfidf_df[features]

# Target
y = tfidf_df['is_serious']

# Instantiate the SVD
svd = TruncatedSVD(n_components=500)

# Fit and transform
svd_matrix = svd.fit_transform(X)

print('SVD Matrix Shape:    ', svd_matrix.shape)
print('Before SVD Shape:    ', cv_df.shape)
print('SVD Components Shape:', svd.components_.shape)

# How many components do we need to represent at least %70 of the variance in the data?
np.where(np.cumsum(svd.explained_variance_ratio_) < .8)[0][-1]+1

# Coonvert to dataframe
component_names = ["component_"+str(i+1) for i in range(100)]
svd_df = pd.DataFrame(svd_matrix,
                      columns=component_names)

svd_df.head()

# Try on the Logistic Regression model
X_train, X_test, y_train, y_test = train_test_split(svd_df, y, stratify=y, random_state=42)

lr = LogisticRegression(C = 10, max_iter = 150, solver = 'liblinear')
lr.fit(X_train, y_train)

print('Cross Val Score: ', cross_val_score(lr, X_train, y_train, cv = 5).mean())
print('Train Score:     ', lr.score(X_train, y_train))
print('Test Score:      ', lr.score(X_test, y_test))

Observations: The cross val score is very similar but our training and test score are much closer than previously (Train: 0.9992125984251968, Test: 0.8301886792452831) which means that the model is less overfit

### Final Model: Logistic Regression with Tf-Idf and SVD

# Try on the Logistic Regression model
X_train, X_test, y_train, y_test = train_test_split(svd_df, y, stratify=y, random_state=42)

lr = LogisticRegression(C = 10, max_iter = 150, solver = 'liblinear')
lr.fit(X_train, y_train)

# Score the model/Evaluate Model
print('--------------MODEL EVALUTATION---------------')
print('Cross Val Score: ', cross_val_score(lr, X_train, y_train, cv = 5).mean())
print('Train Score:     ', lr.score(X_train, y_train))
print('Test Score:      ', lr.score(X_test, y_test))
print()
 
# Get predictions
predictions = lr.predict(X_test)
    
# Confusion matrix and measures
confusion_matrix_results(y_test, predictions)

Observation: 
- The model has false predictions pretty evenly
- Almost the same number of false positives and negatives
- Sensativity and specificty are pretty close
