==> code/07_Naive_Bayes.ipynb <==
# Project 3: Naive Bayes Model

# Imports
import pandas as pd
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import confusion_matrix

# Read in Count Vectorize Data
cv_df = pd.read_csv('../data/count_vec.csv')
tfidf_df = pd.read_csv('../data/tfidf.csv')

cv_df.head()

tfidf_df.head()

### Baseline Score

# Check for unbalanced classes
cv_df['is_serious'].value_counts(normalize=True)
# Do not need to worry about unbalanced classes given the percentages

### Naive Bayes Model (Count Vectorize)

# Set X and y
features = [column for column in cv_df.columns if column != 'is_serious']
features.remove('sent_compound')
features.remove('sent_neu')
X = cv_df[features]
y = cv_df['is_serious']

X.head()

# Train, test split
X_train, X_test, y_train, y_test = train_test_split(X,
                                                    y,
                                                    random_state=42,
                                                    stratify=y)

# Instantiate the model
nb = MultinomialNB()

# Fit the model
nb.fit(X_train, y_train)

# Score the model
# Score the model
print('--------------MODEL EVALUTATION---------------\n')
print('Accuracy:     ', cross_val_score(nb, X_train, y_train, cv = 5).mean())
print('Train Score:  ', nb.score(X_train, y_train))
print('Test Score:   ', nb.score(X_test, y_test))
    
# Get predictions
predictions = nb.predict(X_test)
    
# Confusion Matrixx
cm = confusion_matrix(y_test, predictions)
print('\n-------------CONFUSION MATRIX---------------\n')
print(pd.DataFrame(cm, 
                   columns=['pred neg', 'pred pos'], 
                   index = ['actual neg', 'actual pos']))
    

# Results
tn, fp, fn, tp = confusion_matrix(y_test, predictions).ravel()
print("\nTrue Negatives: %s" % tn)
print("False Positives: %s" % fp)
print("False Negatives: %s" % fn)
print("True Positives: %s" % tp)
print('\n------------------METRICS-------------------\n')
print(f'Sensativity: {round(tp / (tp + fn),4)}')
print(f'Specificity: {round(tn / (tn + fp),4)}')

Observations:
- This model looks pretty good...showing almost as good of metrics as our Logistic Regression model
- Good sensativity and specificity (do not necessarily need to optimize for one over the other since a false postiive is no worse than a false negative)
    - Although...in the case you were using this to flag posts where users are mentioning serious topics such as suicide you would not want to incorrectly mark a serious post as a casual post (false negatives)

### Naive Bayes Model (TF-IDF)

# Set X and y
features = [column for column in tfidf_df.columns if column != 'is_serious']
features.remove('sent_compound')
features.remove('sent_neu')
X = cv_df[features]
y = cv_df['is_serious']

# Train, test split
X_train, X_test, y_train, y_test = train_test_split(X,
                                                    y,
                                                    random_state=42,
                                                    stratify=y)

# Instantiate the model
nb = MultinomialNB()

# Fit the model
nb.fit(X_train, y_train)

# Score the model
# Score the model
print('--------------MODEL EVALUTATION---------------\n')
print('Accuracy:     ', cross_val_score(nb, X_train, y_train, cv = 5).mean())
print('Train Score:  ', nb.score(X_train, y_train))
print('Test Score:   ', nb.score(X_test, y_test))
    
# Get predictions
predictions = nb.predict(X_test)
    
# Confusion Matrixx
cm = confusion_matrix(y_test, predictions)
print('\n-------------CONFUSION MATRIX---------------\n')
print(pd.DataFrame(cm, 
                   columns=['pred neg', 'pred pos'], 
                   index = ['actual neg', 'actual pos']))
    

# Results
tn, fp, fn, tp = confusion_matrix(y_test, predictions).ravel()
print("\nTrue Negatives: %s" % tn)
print("False Positives: %s" % fp)
print("False Negatives: %s" % fn)
print("True Positives: %s" % tp)
print('\n------------------METRICS-------------------\n')
print(f'Sensativity: {round(tp / (tp + fn),4)}')
print(f'Specificity: {round(tn / (tn + fp),4)}')

Observations: this model did only slightly worse than the model with count vectorize. I will stick with count vectorize

### GridSearch (Naive Bayes w/ Count Vectorize)
- With gridsearch this model performs better than TF-IDF

# Set X and y
features = [column for column in cv_df.columns if column != 'is_serious']
features.remove('sent_compound')
features.remove('sent_neu')
X = cv_df[features]
y = cv_df['is_serious']

# Train, test split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42, stratify=y)

# Instantiate Gridsearch.
grid = GridSearchCV(estimator = MultinomialNB(),
                    param_grid = {
                        'alpha': [0.4, 0.5, 0.6, 0.65, 0.7]},
                    cv = 5)

grid.fit(X_train, y_train)

# Best parameters & score
print('Best Params: ', grid.best_params_)

# What is the best score?
print('Best Score: ', grid.best_score_)

# Test model with new parameters

# Instantiate the model
nb = MultinomialNB(alpha = 0.4)

# Fit the model
nb.fit(X_train, y_train)

# Score the model
# Score the model
print('--------------MODEL EVALUTATION---------------\n')
print('Accuracy:     ', cross_val_score(nb, X_train, y_train, cv = 5).mean())
print('Train Score:  ', nb.score(X_train, y_train))
print('Test Score:   ', nb.score(X_test, y_test))
    
# Get predictions
predictions = nb.predict(X_test)
    
# Confusion Matrixx
cm = confusion_matrix(y_test, predictions)
print('\n-------------CONFUSION MATRIX---------------\n')
print(pd.DataFrame(cm, 
                   columns=['pred neg', 'pred pos'], 
                   index = ['actual neg', 'actual pos']))
    

# Results
tn, fp, fn, tp = confusion_matrix(y_test, predictions).ravel()
print("\nTrue Negatives: %s" % tn)
print("False Positives: %s" % fp)
print("False Negatives: %s" % fn)
print("True Positives: %s" % tp)
print('\n------------------METRICS-------------------\n')
print(f'Sensativity: {round(tp / (tp + fn),4)}')
print(f'Specificity: {round(tn / (tn + fp),4)}')
